refactor: rename @kampus/tailwind-config to @kampus/tailwind
