+++ apps/kampus/package.json
@@ -10,7 +10,7 @@
     "relay": "relay-compiler"
   },
   "devDependencies": {
-    "@kampus/tailwind-config": "*",
+    "@kampus/tailwind": "*",
     "@types/react-relay": "14.1.4",
     "relay-compiler": "15.0.0"
   },
