+++ apps/kampus/postcss.config.js
@@ -1,2 +1,2 @@
 // @ts-expect-error - No types for postcss
-module.exports = require("@kampus/tailwind-config/postcss");
+module.exports = require("@kampus/tailwind/postcss");
