+++ apps/ui/package.json
@@ -13,7 +13,7 @@
   "license": "ISC",
   "devDependencies": {
     "@kampus/kampus": "*",
-    "@kampus/tailwind-config": "*",
+    "@kampus/tailwind": "*",
     "@storybook/addon-a11y": "^7.0.20",
     "@storybook/addon-essentials": "^7.0.20",
     "@storybook/addon-interactions": "^7.0.20",
