+++ packages/tailwind-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kampus/tailwind-config",
+  "name": "@kampus/tailwind",
   "version": "0.0.0",
   "description": "",
   "main": "index.js",
