+++ packages/ui/package.json
@@ -10,7 +10,7 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@kampus/tailwind-config": "0.0.0"
+    "@kampus/tailwind": "0.0.0"
   },
   "dependencies": {
     "@radix-ui/react-alert-dialog": "^1.0.4",
